requirements.txt
streamlit
openai
google-analytics-data
google-ads
python-dotenv
sqlalchemy
psycopg2-binary
pandas
google-auth-oauthlib
google-api-python-client
plotly
requests